Add video and audio quality flags (#42)
* Add --video-quality and --audio-quality flags for videos served over DMC

User can specify `highest`, `lowest`, or a quality tag. If the user specifies anything less than the highest qualities for video and audio, the overall quality will be knocked to `low`. If a user requests an unavailable quality, the downloader raises FormatNotAvailableException and lists available qualities.

* Fix quality determination, use more lenient overwriting on resume

DMC videos report an overall quality of `auto` when using the highest sources, otherwise `low`. Legacy videos report an overall quality of `auto` or `low` depending on what is served. Flash videos always report an overall quality of `auto`. As this opens up many new quality combinations, the downloader no longer checks against quality when resuming. If the byte position of the existing files exceed the file to be downloaded, the downloader raises FormatNotAvailableException. If the byte comparison block does not match or exceeds the length of the existing file, the existing file will be deleted.