Update video response parameters collection
Fixes #94. As of 2021, the Flash videos page was discontinued and the HTML5 player parameters schema was updated to accomodate all video types. All videos now appear to be served exclusively over DMC but we may need to add legacy support back as we look at all videos. Quality detection needs to be fixed.